+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## Release 4.1.0
+
+### New features
+- Added event based usage of the handleOnNewImageProcessingScores function
+
+### Known issues
+- handleOnNewImageProcessingScores crashes on SIM1012 with firmware versions other than [2.1.0, 2.2.0, 2.2.1]
+
 ## Release 4.0.0
 
 ### Improvements